+++ src/Admin.fs
@@ -11,12 +11,11 @@ module Admin =
             )
         )
 
-    let createAdminFromConsumer configuration groupId =
-        use consumer = Consumer.createConsumer configuration.BrokerList configuration.Topic groupId
-        new AdminClient(consumer.Handle)
+    let createAdminFromHandle (handle: Handle) =
+        new AdminClient(handle)
 
     let getAllTopics (admin: AdminClient) =
-        admin.GetMetadata(TimeSpan.FromSeconds 10.0)
+        admin.GetMetadata(TimeSpan.FromSeconds 5.0)
         |> fun metadata ->
             metadata.Topics
             |> Seq.choose (fun topic ->
@@ -30,3 +29,13 @@ module Admin =
         admin
         |> getAllTopics
         |> List.contains topic
+
+    let isUp (admin: AdminClient) =
+        try
+            admin.GetMetadata(TimeSpan.FromSeconds 5.0)
+            |> fun metadata ->
+                metadata.Brokers
+                |> Seq.isEmpty
+                |> not
+        with
+        | :? KafkaException -> false