+++ src/Types.fs
@@ -4,6 +4,8 @@ namespace Kafka
 // Common
 //
 
+[<Measure>] type second
+
 type BrokerList = BrokerList of string
 type StreamName = StreamName of string
 
@@ -23,24 +25,16 @@ module GroupId =
         | Id groupId -> groupId
         | Random -> sprintf "random-%d" System.DateTime.Now.Ticks    // random (unique) group id means, it will always starts from the beginning
 
-type Configuration = {
+type ServiceStatus = {
+    MarkAsEnabled: unit -> unit
+    MarkAsDisabled: unit -> unit
+}
+
+type Logger = {
+    Log: string -> unit
+}
+
+type ConnectionConfiguration = {
     BrokerList: BrokerList
     Topic: StreamName
 }
-
-//
-// Kafka readers
-//
-
-type DecodedMessageReader = {
-    ReadMessage: string -> unit
-}
-
-type ParsedMessageReader<'Event> = {
-    ParseEvent : string -> 'Event
-    OnEvent: 'Event -> unit
-}
-
-type MessageReader<'Event> =
-    | DecodedMessageReader of DecodedMessageReader
-    | ParsedMessageReader of ParsedMessageReader<'Event>